backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-enumfields==2.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
sqlparse==0.5.3
types-PyYAML==6.0.12.20241221
typing_extensions==4.12.2
